pyproject.toml
[project]
name = "go-easy-your-report"
version = "0.1.0"
description = "Document-first report workflow scripts"
requires-python = ">=3.11"
dependencies = [
  "Pygments>=2.18.0,<3",
  "python-docx>=1.2.0,<2",
]

[tool.uv]
package = false
